+++ Classes/Tca/ContentBehaviour.php
@@ -120,7 +120,7 @@ class ContentBehaviour extends AbstractContentRow
         $setValue = htmlspecialchars(implode(',', $setClasses));
         // Allow admins to see the internal identifiers
         $inputType = 'hidden';
-        if ($this->isAdmin()) {
+        if ($this->isAdminAndDebug()) {
             $inputType = 'text';
         }
         // Build hidden field structure