+++ Classes/Tca/ContentEnforceEqualColumnHeight.php
@@ -105,7 +105,7 @@ class ContentEnforceEqualColumnHeight extends AbstractContentRow
         $setValue = htmlspecialchars(implode(',', $setClasses));
         // Allow admins to see the internal identifiers
         $inputType = 'hidden';
-        if ($this->isAdmin()) {
+        if ($this->isAdminAndDebug()) {
             $inputType = 'text';
         }
         // Build hidden field structure