+++ Classes/Tca/ContentResponsive.php
@@ -217,7 +217,7 @@ class ContentResponsive extends AbstractContentRow
         $setValue = htmlspecialchars(implode(',', $setClasses));
         // Allow admins to see the internal identifiers
         $inputType = 'hidden';
-        if ($this->isAdmin()) {
+        if ($this->isAdminAndDebug()) {
             $inputType = 'text';
         }